Add logo to nav bar and move to common components.
